requirements.txt
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
llama_index==0.10.37
python-dotenv==1.0.1
streamlit==1.31.1
